updated hx711 library

+++ lib/HX711/HX711.h
@@ -19,14 +19,13 @@
 class HX711
 {
 	private:
-		uint8_t PD_SCK;	    // Power Down and Serial Clock Input Pin
-		uint8_t DOUT;		    // Serial Data Output Pin
-		uint8_t GAIN;	    	// amplification factor
-		long OFFSET = 0.00;	// used for tare weight
-		float SCALE = 1;  	// used to return weight in grams, kg, ounces, whatever
-
-    void _insertSort(double * array, uint8_t size);
-
+		byte PD_SCK;	// Power Down and Serial Clock Input Pin
+		byte DOUT;		// Serial Data Output Pin
+		byte GAIN;		// amplification factor
+		long OFFSET = 0;	// used for tare weight
+		float SCALE = 1;	// used to return weight in grams, kg, ounces, whatever
+		
+    	void _insertSort(double * array, uint8_t size);
 	public:
 
 		HX711();
@@ -38,7 +37,7 @@ class HX711
 		// - With a gain factor of 64 or 128, channel A is selected
 		// - With a gain factor of 32, channel B is selected
 		// The library default is "128" (Channel A).
-		void begin(uint8_t dout, uint8_t pd_sck, uint8_t gain = 128);
+		void begin(byte dout, byte pd_sck, byte gain = 128);
 
 		// Check if HX711 is ready
 		// from the datasheet: When output data is not ready for retrieval, digital output pin DOUT is high. Serial clock
@@ -53,13 +52,13 @@ class HX711
 		// set the gain factor; takes effect only after a call to read()
 		// channel A can be set for a 128 or 64 gain; channel B has a fixed 32 gain
 		// depending on the parameter, the channel is also set to either A or B
-		void set_gain(uint8_t gain = 128);
+		void set_gain(byte gain = 128);
 
 		// waits for the chip to be ready and returns a reading
 		long read();
 
 		// returns an average reading; times = how many times to read
-		double read_average(uint8_t times = 10);
+		long read_average(byte times = 10);
 		
 		// returns the median reading; times = how many times to read
 		double read_median(uint8_t times = 7);
@@ -67,21 +66,21 @@ class HX711
 		// return the maximum reading; times = how many times to read
 		double read_max(uint8_t times = 10);
 
-    // returns (read_median() - OFFSET), that is the current median value without the tare weight; times = how many readings to do
-    double get_median_value(uint8_t times);
+		// returns (read_median() - OFFSET), that is the current median value without the tare weight; times = how many readings to do
+		double get_median_value(uint8_t times);
 
-    // returns (read_max() - OFFSET), that is the current max value without the tare weight; times = how many readings to do
-    double get_max_value(uint8_t times);
+		// returns (read_max() - OFFSET), that is the current max value without the tare weight; times = how many readings to do
+		double get_max_value(uint8_t times);
 
 		// returns (read_average() - OFFSET), that is the current value without the tare weight; times = how many readings to do
-		double get_value(uint8_t times = 1);
+		double get_value(byte times = 1);
 
 		// returns get_value() divided by SCALE, that is the raw value divided by a value obtained via calibration
 		// times = how many readings to do
-		float get_units(uint8_t times = 1);
+		float get_units(byte times = 1);
 
 		// set the OFFSET value for tare weight; times = how many times to read the tare value
-		void tare(uint8_t times = 10);
+		void tare(byte times = 10);
 
 		// set the SCALE value; this value is used to convert the raw data to "human readable" data (measure units)
 		void set_scale(float scale = 1.f);
